// src/runnersQcApp/containers/CreateLocationContainer/types.ts
export enum IdType {
  AL = "AL",
  AK = "AK",
  AZ = "AZ",
  AR = "AR",
  CA = "CA",
  CO = "CO",
  CT = "CT",
  DE = "DE",
  FL = "FL",
  GA = "GA",
  HI = "HI",
  ID = "ID",
  IL = "IL",
  IN = "IN",
  IA = "IA",
  KS = "KS",
  KY = "KY",
  LA = "LA",
  ME = "ME",
  MD = "MD",
  MA = "MA",
  MI = "MI",
  MN = "MN",
  MS = "MS",
  MO = "MO",
  MT = "MT",
  NE = "NE",
  NV = "NV",
  NH = "NH",
  NJ = "NJ",
  NM = "NM",
  NY = "NY",
  NC = "NC",
  ND = "ND",
  OH = "OH",
  OK = "OK",
  OR = "OR",
  PA = "PA",
  RI = "RI",
  SC = "SC",
  SD = "SD",
  TN = "TN",
  TX = "TX",
  UT = "UT",
  VT = "VT",
  VA = "VA",
  WA = "WA",
  WV = "WV",
  WI = "WI",
  WY = "WY",
}

export enum ValueType {
  Alabama = "Alabama, AL",
  Alaska = "Alaska, AK",
  Arizona = "Arizona, AZ",
  Arkansas = "Arkansas, AR",
  California = "California, CA",
  Colorado = "Colorado, CO",
  Connecticut = "Connecticut, CT",
  Delaware = "Delaware, DE",
  Florida = "Florida, FL",
  Georgia = "Georgia, GA",
  Hawaii = "Hawaii, HI",
  Idaho = "Idaho, ID",
  Illinois = "Illinois, IL",
  Indiana = "Indiana, IN",
  Iowa = "Iowa, IA",
  Kansas = "Kansas, KS",
  Kentucky = "Kentucky, KY",
  Louisiana = "Louisiana, LA",
  Maine = "Maine, ME",
  Maryland = "Maryland, MD",
  Massachusetts = "Massachusetts, MA",
  Michigan = "Michigan, MI",
  Minnesota = "Minnesota, MN",
  Mississippi = "Mississippi, MS",
  Missouri = "Missouri, MO",
  Montana = "Montana, MT",
  Nebraska = "Nebraska, NE",
  Nevada = "Nevada, NV",
  NewHampshire = "New Hampshire, NH",
  NewJersey = "New Jersey, NJ",
  NewMexico = "New Mexico, NM",
  NewYork = "New York, NY",
  NorthCarolina = "North Carolina, NC",
  NorthDakota = "North Dakota, ND",
  Ohio = "Ohio, OH",
  Oklahoma = "Oklahoma, OK",
  Oregon = "Oregon, OR",
  Pennsylvania = "Pennsylvania, PA",
  RhodeIsland = "Rhode Island, RI",
  SouthCarolina = "South Carolina, SC",
  SouthDakota = "South Dakota, SD",
  Tennessee = "Tennessee, TN",
  Texas = "Texas, TX",
  Utah = "Utah, UT",
  Vermont = "Vermont, VT",
  Virginia = "Virginia, VA",
  Washington = "Washington, WA",
  WestVirginia = "West Virginia, WV",
  Wisconsin = "Wisconsin, WI",
  Wyoming = "Wyoming, WY",
}

export type TOption = {
  value: string;
  label: string;
};

export enum CountryType {
  USA = "USA",
  Canada = "Canada",
  Mexico = "Mexico",
  UK = "UK",
  France = "France",
  Germany = "Germany",
  Ireland = "Ireland",
  Australia = "Australia",
}